Started refactoring of basic code and functionality
Separated code into main.c (with main function), game.h (declarations
and function signatures), and game.c (defining a lot of variables and
functions to get them out of the way a bit). This will hopefully greatly
reduce the cluttered SINGLE FILE I have been using so far, and it should
be easier to read. I hope. If you are reading this, and are NOT me, I
pray for your sanity.